% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{agl_mean}
\alias{agl_mean}
\title{Angle average}
\usage{
agl_mean(x)
}
\arguments{
\item{x}{angle in radians.}
}
\description{
Angle average
}
\examples{
agl_mean(c(-pi, pi))
}
\keyword{angle}
\keyword{internal}

