% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{agl_rescale}
\alias{agl_rescale}
\title{Rescale angle to [-pi; pi]}
\usage{
agl_rescale(x)
}
\arguments{
\item{x}{angle in radians.}
}
\description{
Rescale angle to [-pi; pi]
}
\examples{
agl_rescale(5*pi / c(-4, 4))
}
\keyword{angle}
\keyword{internal}

