% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{as.data.frame.bsm}
\alias{as.data.frame.bsm}
\title{Coerce brokenstick model to data.frame}
\usage{
\method{as.data.frame}{bsm}(x, row.names = NULL, ...)
}
\arguments{
\item{x}{a brokenstick model.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Coerce brokenstick model to data.frame
}
\examples{
data(exses)
bsm <- tdrply(brokenstick, 1:2, no = 50:53, obj = exses)
lapply(bsm, as.data.frame)
}
\keyword{brokenstick}
\keyword{internal}

