% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_class.r
\name{as.ses}
\alias{as.ses}
\alias{as.stat}
\alias{as.tdr}
\alias{is.ses}
\alias{is.stat}
\alias{is.tdr}
\title{Creates or tests for 'ses'.}
\usage{
as.ses(...)

is.ses(x)

is.tdr(x)

is.stat(x)

as.tdr(x)

as.stat(x)
}
\arguments{
\item{...}{Elements to group into an 'ses' object.}

\item{x}{A object to test.}
}
\description{
Creates or tests for 'ses'.
}
\details{
\code{as.ses} coerces an object to class \code{ses}. An \code{ses}
object should contain:
\itemize{
  \item name The name of the individual.
  \item hash The md5sum hash of the original dataset.
  \item tdr The TDR data.
  \item stat The dives statistics.
  \item delim A \code{data.frame} that stores dive and bottom delimitation results.
}

\code{as.ses} tests if an object belong to the \code{'ses'} class

\code{is.tdr} Test if an object belong to the \code{'tdr'} class

\code{is.stat} tests if an object belong to the \code{'statdives'} class

\code{as.tdr} coerces an object to class \code{tdr}. A TDR table should
contain at least columns time and depth in that order.

\code{as.stat} coerces an object to class \code{stat}.
}
\examples{
new.ses <- list(name = "Example", tdr = data.frame(), stat = data.frame())
is.ses(new.ses)
is.stat(new.ses$stat)

new.ses <- as.ses(new.ses)
is.ses(new.ses)
is.stat(new.ses$stat)
}
\keyword{internal}

