% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{at_ker}
\alias{at_ker}
\title{Test if animal is located around Kerguelen}
\usage{
at_ker(lon, lat, r = 130)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{r}{radius}
}
\description{
Test if animal is located around Kerguelen
}
\details{
Kerguelen location taken at (49.353282 deg. S,69.354630 deg. E)
}
\examples{
data(exses)
is_at_kerguelen <- at_ker(exses$stat$lon, exses$stat$lat)
}

