% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_plot.r
\name{axisPOSIXct}
\alias{axisPOSIXct}
\title{Draw a correct axis for POSIXct objects}
\usage{
axisPOSIXct(x, side = 1, format = "\%m/\%d")
}
\arguments{
\item{x}{The POSIXct data.}

\item{side}{an integer specifying which side of the plot the axis is
    to be drawn on.  The axis is placed as follows: 1=below,
    2=left, 3=above and 4=right.}

\item{format}{The format for ticks label.}
}
\description{
Draw a correct axis for POSIXct objects
}
\keyword{internal}

