% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_biopic.r
\name{biolum_events}
\alias{biolum_events}
\title{Identify potential bioluminescence emission events}
\usage{
biolum_events(x, sensitivity = 1.26, max_light = exp(-20))
}
\arguments{
\item{x}{a data.frame such as returned by \code{\link{BioPIC}} output}

\item{sensitivity}{A threshold to decide if a signal peak is high enought. See
details.}

\item{max_light}{Max light level (W/cm^2) where a event can be detected. Set to
NULL to disable.}
}
\description{
Identify potential bioluminescence emission events
}
\details{
Sensor sensitivity threshold: The minimum ratio between two measures
to be considered significantly different. In laboratory experiments, the
sensors measured light intensity +- 2 units while submitted to a constant
light intensity. Translating the sensor log-scale units to SI linear scale
units this accuracy measure translates into a minimum ratio of 1.26
}
\examples{
\dontrun{
data(exses)
exses$tdr$light_si <- SI_light(exses$tdr$light)
biolum_tbl <- tdrply(BioPIC, obj = exses)
ble_tbl <- Reduce(rbind, lapply(biolum_tbl, biolum_events))
}
}
\keyword{light}

