% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_biopic.r
\name{biopic.qr}
\alias{biopic.qr}
\title{BioPIC QR decomposition on a TDR sample}
\usage{
biopic.qr(x, lightSI.nm = "light_si")
}
\arguments{
\item{x}{A data subset of fixed width of 11 seconds/lines.}

\item{lightSI.nm}{The name of the TDR column with light values in W/cm^2.}
}
\description{
BioPIC QR decomposition on a TDR sample
}
\references{
Vacquié-Garcia, J., Royer, F., Dragon, A.-C., Viviant, M., Bailleul, F. 
& Guinet, C. (2012) Foraging in the Darkness of the Southern Ocean: 
Influence of Bioluminescence on a Deep Diving Predator. PLoS ONE, 7, e43565.
}
\keyword{internal}

