% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_biopic.r
\name{BioPIC}
\alias{BioPIC}
\title{BioPIC: bioluminescent event detection tool}
\usage{
BioPIC(x, lightSI.nm = "light_si", nms = c("alpha", "lessK", "Ia"))
}
\arguments{
\item{x}{a TDR data sample}

\item{lightSI.nm}{The name of the TDR column with light values in W/cm^2.}

\item{nms}{The names of the output variables.}
}
\description{
Adaptation of BioPIC method with 1 Hz sampling frequency datasets and 
of post-peak attenuation (related to sensor and not depth)
}
\examples{
data(exses)
dv <- tdrply(identity, no = 100, obj = exses)[[1]]
dv$light_si <- SI_light(dv$light)
biolum_tbl <- BioPIC(dv)
}
\references{
Vacquie-Garcia, J., Royer, F., Dragon, A.-C., Viviant, M., Bailleul, F. 
& Guinet, C. (2012) Foraging in the Darkness of the Southern Ocean: 
Influence of Bioluminescence on a Deep Diving Predator. PLoS ONE, 7, e43565.
}
\keyword{light}

