% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_delim.r
\name{bottom_delim_bsm}
\alias{bottom_delim_bsm}
\title{Delimitate bottom phase of dive using brokenstick models}
\usage{
bottom_delim_bsm(time, depth = NULL, npts = 6)
}
\arguments{
\item{time}{time readings, sorted in chronological order.}

\item{depth}{depth readings, sorted in chronological order.}

\item{pts}{The number of points to use in the brokenstick model. The default
(\code{npts = 6}) is used by CTD-SRDL tags (Sea Mammal Research unit, St Andrews
University) to summarize dive profiles.}
}
\description{
The bottom is defined as the dive period between the first brokenstick
segment (descent) and the last brokenstick segment (ascent).
}
\examples{
data(exses)
ind(exses)

n <- 65
idx <- tdrply(bottom_delim_bsm, 1:2, "!_/", no = n)[[1]]
tdrply(plot, 1:2, "!_/", no = n, main = n)
tdrply(function(x, st, ed, ...) points(x[st:ed, ]), 1:2, no = n, la = idx)
}
\keyword{internal}

