% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_delim.r
\name{bottom_delim_std}
\alias{bottom_delim_std}
\title{Delimitate bottom phase of dive using depth threshold}
\usage{
bottom_delim_std(depth, ledge = 0.8)
}
\arguments{
\item{depth}{depth readings, sorted in chronological order.}

\item{ledge}{depth threshold, specified as a percentage which is compared
to the maximum depth.}
}
\description{
Classic bottom delimitation method that defines bottom as the period between
the first and last passage at a depth specified as a percentage of the
maximum depth in the dive.
}
\examples{
data(exses)
ind(exses)

n <- 65
idx <- tdrply(bottom_delim_std, 2, "!_/", no = n)[[1]]
tdrply(plot, 1:2, "!_/", no = n, main = n)
tdrply(function(x, st, ed) points(x[st:ed, ]), 1:2, no = n, la = idx)
}
\keyword{internal}

