% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_delim.r
\name{bottom_delim_vspd}
\alias{bottom_delim_vspd}
\title{Delimitate bottom phase of dive using vertical speed threshold}
\usage{
bottom_delim_vspd(time = NULL, y = NULL, y.type = c("depth", "vspd"),
  vert_vel = 0.75, w = 21)
}
\arguments{
\item{time}{time readings, sorted in chronological order.}

\item{vert_vel}{the verical speed threshold.}

\item{w}{a window width to use for vertical speed smoothing (moving average) 
when \code{vspd} is not provided.}

\item{depth}{depth readings, sorted in chronological order.}

\item{vspd}{smoothed verical speed, sorted in chronological order. Optional. 
If provided, function is faster as it does not need to compute it from 
depth and time.}
}
\description{
Delimitate bottom which is defined as the period between the first and last 
moment in the dive were vertical speed (simplified signal using \code{\link{poly}}) 
exceeds a threshold.
}
\examples{
data(exses)
ind(exses)

n <- 65
idx <- tdrply(bottom_delim_vspd, 1:2, "!_/", no = n)[[1]]
tdrply(plot, 1:2, "!_/", no = n, main = n)
tdrply(function(x, st, ed) points(x[st:ed, ]), 1:2, no = n, la = idx)
}
\keyword{internal}

