% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_behavior.r
\name{broadness_index}
\alias{broadness_index}
\title{Broadness index}
\usage{
broadness_index(x = ind(), time_col = 1)
}
\arguments{
\item{x}{input data, a \code{ses} object.}

\item{time_col}{Character or numeric giving the column of the TDR table that
stores the timestamps.}
}
\description{
The duration of the bottom phase divided by the duration of the dive
}
\details{
The index ranges from 0 (short bottom) to 1 (long bottom).
}
\examples{
data(exses)
exses$stat$brd <- broadness_index(exses)

plot(brd ~ time, exses$stat)
}
\references{
Halsey, L.G., Bost, C.-A., Handrich, Y. (2007) A thorough and
quantified method for classifying seabird diving behaviour.
Polar Biology, 30, 991-1004.
}
\keyword{behavior}

