% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{brokenstick}
\alias{brokenstick}
\title{Fitting brokenstick models}
\usage{
brokenstick(x, y, npts = 6, start = NULL, na.action, ...)
}
\arguments{
\item{x}{The x data. Note that the \code{x} values have to be sorted. Can also
be a list of \code{x} and \code{y} (processed by \code{\link{xy.coords}}).
Alternatively, it can also be an object of class "\code{\link{formula}}".}

\item{y}{The y data. (processed by \code{\link{xy.coords}}). Alternatively,
if \code{x} is a \code{'formula'}, \code{y} can also be an optional data frame,
list or environment (or object coercible by \code{\link{as.data.frame}} to a
data frame) containing the variables in the model. If not found in data, the
variables are taken from environment (\code{formula}), typically the environment
from which \code{brokenstick} is called.}

\item{npts}{The number of points for the brokenstick model to fit. See
\code{\link{optBrokenstick}} for a version of \code{brokenstick} which figures
out this number of points automatically.}

\item{start}{Some starting points to start the algorithm with.}

\item{na.action}{A function which indicates what should happen when the data
contain \code{NAs}. The default is set by the \code{na.action} setting of
\code{\link{options}}, and is \code{\link{na.fail}} if that is unset. The
"factory-fresh" default is \code{\link{na.omit}}. Another possible value
is \code{NULL}, no action. Value \code{\link{na.exclude}} can be useful.}

\item{...}{Further arguments to be passed to \code{\link{bsmfit}}
such as \code{eco.mem}.}
}
\value{
A \code{bsm} object with (depending on \code{eco.mem}) contains:
\itemize{
  \item pts.x (\code{eco.mem < 7}) The x values of the brokenstick points.
  \item slope (\code{eco.mem < 6}) The slopes of each "stick" of the model.
  \item intercept (\code{eco.mem < 5}) The intercept of each "stick" of the model.
  \item pts (\code{eco.mem < 4}) The index of the brokenstick points.
  \item na.action Information from the action which was applied to object if \code{NAs}
  were handled specially. (\code{eco.mem} inefficient)
  \item residuals (\code{eco.mem < 3}) The model's residuals.
  \item fitted.values (\code{eco.mem < 2}) The fitted values.
  \item data (\code{eco.mem < 1}) The input data used for fitting.
  \item pts.no The iteration number of points. (\code{eco.mem} inefficient)
}
}
\description{
\code{brokenstick} is used to fit brokenstick models on two-dimensional data
such as Time-Depth, Depth-Temperature or Depth-Light profiles. Brokenstick models
are useful for compressing or extracting the shape of high resolution profiles.
}
\details{
See \code{\link{bsmfit}}, the function called by \code{brokenstick}
on each iteration to fit the model with specified points.
}
\examples{
data(exses)
dv <- tdrply(identity, 1:2, no = 400, obj = exses)[[1]]

# Syntax
bsm <- brokenstick(dv$time, dv$depth)
bsm <- brokenstick(dv) # if two columns
bsm <- brokenstick(depth ~ time, dv)
bsm <- with(dv, brokenstick(depth ~ time))
}
\seealso{
\code{\link{optBrokenstick}} and \code{\link{predict.bsm}}, \code{\link{residuals.bsm}},
\code{\link{update.bsm}}, \code{\link{summary.bsm}},
\code{\link{coef.bsm}}, \code{\link{plot.bsm}}, \code{\link{as.data.frame.bsm}}
for other functions with a S3 method for \code{bsm} objects.
}
\keyword{brokenstick}

