% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{bsmfit}
\alias{bsmfit}
\title{Fitter function for brokenstick models}
\usage{
bsmfit(xy, pts, eco.mem = 0L)
}
\arguments{
\item{xy}{Data, as returned by \code{\link{xy.coords}}. The \code{x} values
have to be sorted.}

\item{pts}{The points to use.}

\item{eco.mem}{An integer value between 0 and 7 to control the memory size of the
output. \code{0} = no memory savings.}
}
\description{
Basic computing engines of the brokenstick algorithm called by
\code{\link{brokenstick}}.
}
\seealso{
\code{\link{brokenstick}} and \code{\link{optBrokenstick}} which should
be used to fit brokenstick models.
}
\keyword{internal}

