% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_tdr.r
\name{check_nomtch}
\alias{check_nomtch}
\title{Deal with missing periods}
\usage{
check_nomtch(x, no_match, obj = ind())
}
\arguments{
\item{x}{a data.frame with "grp" match information.}

\item{no_match}{a string indicating how to handle periods with no match.}

\item{obj}{a "ses" object.}
}
\value{
updated x, warning /error.
}
\description{
\code{check_nomatch} search for periods without entry in "delim" table and
check if a warning/error is to be thrown.
}
\keyword{internal}
\keyword{tdr_extract}

