% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{coef.bsm}
\alias{coef.bsm}
\title{Extract brokenstick models coefficients}
\usage{
\method{coef}{bsm}(object, ...)
}
\arguments{
\item{object}{\code{bsm} object, typically result from \code{\link{brokenstick}}
or \code{\link{optBrokenstick}}.}

\item{...}{other arguments.}
}
\value{
A data frame with slope and intercepts of each stick.
}
\description{
Extract brokenstick models coefficients
}
\details{
Require \code{eco.mem <= 4} (see \code{\link{bsmfit}}).
}
\examples{
data(ses)
dv <- tdrply(identity, 1:2, no = 100, obj= exses)[[1]]
coef(brokenstick(dv))
}
\seealso{
\code{\link{brokenstick}}, \code{\link{optBrokenstick}} for model fitting.
}
\keyword{brokenstick}
\keyword{internal}

