% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functionals.r
\name{compose}
\alias{compose}
\title{Function composition}
\usage{
compose(...)
}
\arguments{
\item{...}{functions to be composed}
}
\value{
\code{compose} returns the composed of functions listed in \code{...}.
Pay attention to the order, compose(f, g) returns \code{g(f())}.
}
\description{
Function composition
}
\details{
see the infix version \code{\link{\%c\%}}
}
\examples{
f <- compose(is.na, sum)
f(c(1, NA, NA))
}
\keyword{functional}

