% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_correction.r
\name{correct_depth}
\alias{correct_depth}
\title{Correct depth sequence from drift of values over time}
\usage{
correct_depth(x, dur = 5000, plt = FALSE, span = 0.15, ...)
}
\arguments{
\item{x}{The depth sequence.}

\item{dur}{A window width for rolling functions. Make sure to choose that 
argument so that the animal went back to the surface at least once in this 
window.}

\item{plt}{Should a plot about processing should be drawn ?}

\item{span}{Argument passed to \code{\link{loess}}.}

\item{...}{Other arguments to be passed to \code{\link{loess}}.}
}
\description{
Correct depth sequence from drift of values over time
}
\examples{
# correct_depth(exses$tdr$depth, plt = TRUE)
}
\keyword{correction}

