% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_correction.r
\name{correct_time}
\alias{correct_time}
\title{Fix jumps and duplicates in TDR time stamps}
\usage{
correct_time(obj, time_seq = 1, verbose = FALSE)
}
\arguments{
\item{obj}{The TDR dataset}

\item{time_seq}{A variable to use in order to subset \code{obj} and extract 
the time column. The time variable has to be in \code{POSIXct} format.}

\item{verbose}{A logical indicating if the function should be verbose.}
}
\description{
Fix jumps and duplicates in TDR time stamps
}
\details{
Can be use for sampling frequency <= 1 Hz.
}
\examples{
data(exses)
exses$tdr <- correct_time(exses$tdr, verbose = TRUE)
}
\keyword{correction}

