% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{delta}
\alias{delta}
\title{Compute difference between extremes in a set of observations}
\usage{
delta(x, ord = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{observations}

\item{ord}{If \code{TRUE}, the output depends on the ranks of observations: 
\code{delta = Last_Extreme - First_Extreme} and therefore returns a negative 
value if \code{max(x)} is enccountered before \code{min(x)} in \code{x} values.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be 
stripped before the computation proceeds.}
}
\description{
Compute difference between extremes in a set of observations
}
\examples{
data(exses)
dives_phases_durations <- tdrply(delta, "time", c("!", "_", "/", "!_/"), obj = exses)
}

