% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_behavior.r
\name{depth_consistency_index}
\alias{depth_consistency_index}
\title{Depth consistency index between two dives}
\usage{
depth_consistency_index(x = ind(), depth_col = 2, fmt = c("vector",
  "matrix"), na.pad = TRUE, ...)
}
\arguments{
\item{x}{input data, a \code{ses} object.}

\item{depth_col}{Character or numeric giving the column of the TDR table that 
stores the depth sequence}

\item{fmt}{Should the function return a vector with pairwise comparison 
of dive taken in chronologic order or a dissimilarity matrix ?}

\item{na.pad}{If TRUE and \code{fmt = "vector"} then a NA is appended to 
the beginning of the result so that the output can be directly allocated 
to the dive statistics table. If TRUE and \code{fmt = "matrix"} then the 
upper triangle of the matrix is filled with NA to save memory.}

\item{...}{Arguments to be passed to \code{\link{depth_range_index}}.}
}
\description{
dri_dive2 * (max_depth_dive2 - max_depth_dive1) / max_depth_dive2
}
\details{
Low values indicate high consistency between the vertical areas 
vistited in the bottom of the two dives (depth range and maximum depth). 
Low values can incitate benthic dives.
}
\examples{
data(exses)
exses$stat$dci <- depth_consistency_index(exses)
hist(exses$stat$dci)
plot(dci ~ time, exses$stat)

Mdci <- depth_consistency_index(exses, fmt = "matrix", na.pad = FALSE)
image(log(Mdci + 0.001)) # Red for values close to 0
}
\references{
Halsey, L.G., Bost, C.-A., Handrich, Y. (2007) A thorough and 
quantified method for classifying seabird diving behaviour. 
Polar Biology, 30, 991-1004.
}
\keyword{behavior}

