% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_behavior.r
\name{depth_range_index}
\alias{depth_range_index}
\title{Depth Range Index}
\usage{
depth_range_index(x = ind(), depth_col = 2, probs = c(0, 1),
  index = TRUE)
}
\arguments{
\item{x}{input data, a \code{ses} object.}

\item{depth_col}{Character or numeric giving the column of the TDR table that 
stores the depth sequence}

\item{probs}{numeric vector of length 2 giving the quantiles probabilities
to be used to compute the range. Default to min and max (\code{c(0, 1)}). 
\code{c(0.1, 0.9)} can be usefull in order to get estimates robust to 
wrong bottom delimitation or unusually high wiggles.}

\item{index}{Should the depth range be divided by the dive maximum depth (index 
ranging between 0 and 1) or the absolute values be returned ?}
}
\description{
Vertical extent of the bottom phase
}
\details{
The index ranges from 0 (perfectly flat bottom) to 1 (large vertical 
width bottom).
}
\examples{
data(exses)
exses$stat$dri <- depth_range_index(exses)

plot(dri ~ time, exses$stat)
}
\references{
Halsey, L.G., Bost, C.-A., Handrich, Y. (2007) A thorough and 
quantified method for classifying seabird diving behaviour. 
Polar Biology, 30, 991-1004.
}
\keyword{behavior}

