% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{df_search}
\alias{df_search}
\title{Search recurssively to a data.frame}
\usage{
df_search(.idx)
}
\arguments{
\item{.idx}{An object.}
}
\description{
This function is a helper designed to be used in \code{tdrply}. It searches
recurssively in an object for a list of data.frames at a given level of depth.
If the search ends to atomic vectors, the function builds a list of
data.frames by taking their elements by two successively.
}
\examples{
.idx <- list(1:10, list(1:10))
# df_search(.idx) # error
.idx <- list(1:10, data.frame(1:10, 1:10))
# df_search(.idx) # error

.idx <- data.frame(1:10, 10:1)
df_search(.idx)
.idx <- list(a = .idx, b = .idx)
df_search(.idx)
.idx <- list(a = 1:3, b = 1:10)
df_search(.idx)
}
\seealso{
\code{\link{tdrply}}
}
\keyword{internal}

