% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_delim.r
\name{dive_delim}
\alias{dive_delim}
\title{Find the dives start and end}
\usage{
dive_delim(obj, thres.dur = c(300, 3000), thres.depth = 15, warn = TRUE,
  ...)
}
\arguments{
\item{obj}{An object of class \code{'ses'}, \code{'statdives'} or \code{'tdr'} 
with time and depth variables named "time" and "depth".}

\item{thres.dur}{The minimum time (seconds) a dive should last to be 
considered as a dive. If \code{thres.dur} has two elements the second is the 
maximum time (seconds) a dive should last to be considered as a dive. If 
\code{thres.dur = NULL} then this threshod is ignored.}

\item{thres.depth}{The depth threshold (in sensor units): diving 
period implies greater values.}

\item{warn}{Should a Warning column be added ?}

\item{...}{Other arguments}
}
\value{
A data frame with the following variable: indice of period start, 
indice of period end, type of period, duration (s), the dive number and a optional 
Warning column. Attributes of the data frame include some information about 
the processing.
}
\description{
Find the dives start and end
}
\examples{
data(exses)
dvs <- dive_delim(exses$tdr)

# Processing info
table(dvs$warning)
attr(dvs, "thres.depth")
attr(dvs, "thres.dur")
}
\seealso{
\code{\link{bottom_delim}}
}

