% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{dive_zone_index}
\alias{dive_zone_index}
\title{Compute goodness of fit of brokenstick models: Dive Zone Index (DZI)}
\usage{
dive_zone_index(x, iter = NULL, n = NULL)
}
\arguments{
\item{x}{a \code{bsm} object as returned by \code{\link{brokenstick}} or
\code{\link{optBrokenstick}}.}

\item{iter}{the iteration number for which the DZI is to be computed. If
\code{iter = NULL} then it is set to last BSM iteration.}

\item{n}{Optional. The number of points to use when calculating the dive zone limits.
if \code{NULL} then \code{n} is set to the number of record in the dataset used to
fit the BSM (when available) or 500 (when not available)}
}
\value{
A \code{dzi} object with:
\itemize{
  \item dzi The dive zone index obtained at each iteration.
  \item max_res The maximum residuals at each iteration.
  \item dz_Lbnd Dive zone lower bound.
  \item dz_Ubnd Dive zone upper bound.
  \item dz_width The difference between the two previous slots i.e. the vertical
  width of the dive zone.
  \item dz_Xval A vector giving the x values corresponding to \code{dz_Lbnd},
  \code{dz_Ubnd} and \code{dz_width}.
  \item no_seg A vector giving the number of BSM segments to which the
  \code{dz_Xval} belong.
  \item seg_width Vertical width covered by BSM segments.
  \item seq_length Duration of BSM segments.
  \item pts.x, pts.y, pts.no Breakpoints information inherited from \code{x}.
  \item data The raw data of input BSM when available.
}
}
\description{
An index of the goodness of fit for brokenstick models. Value of the dive zone
index ranges from 0 (perfect fit) to 1. See original article in references.
}
\details{
If the original TDR data are not available to the function the DZI
of the last BSM iteration will be computed using the maximum residual from the
previous iteration.
}
\examples{
data(exses)
dv <- tdrply(identity, 1:2, no = 100, obj = exses)[[1]]
bsm <- brokenstick(dv)
dzi <- dive_zone_index(bsm)

dzi
str(dzi)
plot(dzi)
}
\references{
Photopoulou, T., Lovell, P., Fedak, M. A., Thomas, L. and Matthiopoulos, J. (2015).
Efficient abstracting of dive profiles using a broken-stick model.
Methods Ecol Evol 6, 278-288. Github repo: https://github.com/theoniphotopoulou/brokenstickmodel.git
}
\keyword{brokenstick}

