% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_behavior.r
\name{entropy}
\alias{entropy}
\title{Shannon entropy index on time at depth proportions}
\usage{
entropy(x, base = 2, scale = FALSE)
}
\arguments{
\item{x}{a character vector naming the depth layers.}

\item{base}{base argument passed to \code{\link{log}}.}

\item{scale}{if TRUE the output is divided by the \code{log(N)} where N is the 
number of layers so that outpout lies between 0 and 1.\code{log(N)} is the 
minimum entropy for sample with N layers where each observation of sample is 
unique i.e all probability are equal.}
}
\description{
Shannon entropy index on time at depth proportions
}
\examples{
data(exses)
ind(exses)
brks <- do.call(seq, as.list(c(range(exses$tdr$depth), by = 2)))
exses$tdr$depth_cat <- as.character(cut(exses$tdr$depth, brks))
plot(tdrply(entropy, "depth_cat", ty = "!_/"), exses$stat$pca)
H <- tdrply(entropy, "depth_cat", ty = "_")
scaled_H <- tdrply(entropy, "depth_cat", ty = "_", scale = TRUE)
plot(H, exses$stat$pca)
plot(H, scaled_H)
}
\keyword{behavior}

