% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_biopic.r
\name{exp10}
\alias{exp10}
\title{Reciprocal function of base::log10}
\usage{
exp10(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Reciprocal function of base::log10
}
\keyword{internal}
\keyword{light}

