% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{exses}
\alias{exses}
\title{An tiny dataset (about 10 days) of an SES equipped with an accelerometer}
\format{An object of class \code{'ses'}.}
\usage{
data(exses)
}
\description{
An tiny dataset (about 10 days) of an SES equipped with an accelerometer
}
\seealso{
\code{\link{as.ses}}
}
\keyword{datasets}

