% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_delim.r
\name{find_drift_roll}
\alias{find_drift_roll}
\title{Find drift dives using roll angle from static accelerometry}
\usage{
find_drift_roll(object, thres.roll = pi/2, thres.dur = 400)
}
\arguments{
\item{object}{a "ses" object}

\item{thres.roll}{threshold to apply to roll in order to discriminate
"on belly" and "on back" attitude}

\item{thres.dur}{minimum duration for a continuous "on back" period to be
considered as a drifting period.}
}
\value{
A data frame one row per drift period with
the start/end indices (refering to TDR rows) of the drift periods,
dive number where the drifting period was observed, the starting time of the
drift and the corresponding drift rate (m/s).
}
\description{
Find drift dives using roll angle from static accelerometry
}
\examples{
data(exses)
tab <- find_drift_roll(exses)
plot(drift_rate ~ time, tab, ylim = c(-0.8, 0.8))
tdrply(plot, 1:2, no = sample(tab$no_dive, 1), obj = exses)
}
\keyword{drift}

