% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_delim.r
\name{find_drift_tdr}
\alias{find_drift_tdr}
\title{Find drift dives using hight-resolution TDR data}
\usage{
find_drift_tdr(object, thres.bsm = 5, bsm = NULL)
}
\arguments{
\item{object}{a \code{SES} object including a TDR dataset, a "dives statistics"
table and a "delim" table.}

\item{thres.bsm}{a threshold to use with \code{\link{max_dist_cost}} when
running \code{\link{optBrokenstick}}.}

\item{bsm}{A list of brokenstick models to use directly instead of computing a
new one with \code{thres.bsm} as parameter.}
}
\description{
\code{find_drift_tdr} takes an object and return a data frame with all the brokenstick
segments along the trip with statistics such as their vertical speed (average
and standard deviation), their duration and so on. This data frame is to be used to
identify the drift segments manually (see examples for an example of method).
}
\examples{
data(exses)
tab <- find_drift_tdr(exses)

\dontrun{
# The main criterion is the duration of segments. Then the R2 and the average vertical speed.
require(manipulate)
manipulate(
{
 cnd <- tab$bsm_r2 >= r2 & tab$dur >= dur * 60 & abs(tab$bsm_slope) <= slope
 cnd <- ifelse(is.na(cnd), FALSE, cnd)
 plot(bsm_slope ~ time, tab[cnd, ], xlim = range(tab$time), ylim = c(-0.8, 0.8))
 abline(h = 0, col = 'gray')
},
 r2 = slider(.95, 1, initial = .99, step = 0.0005, label = "bsm R^2"),
 dur = slider(0, 30, initial = 5, step = 0.25, label = "Duration (minutes)"),
 slope = slider(0, 4, initial = 0.6, step = 0.05, label = "Max. vertical speed (m/s)")
)

# Drift dive example
tdrply(plot, 1:2, no = 300, type = 'l', obj = exses)
}
}
\keyword{drift}

