% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{first}
\alias{first}
\alias{last}
\title{Return first or last element of a list of vector}
\usage{
first(x)

last(x)
}
\arguments{
\item{x}{a list of vector}
}
\description{
Return first or last element of a list of vector
}
\examples{
first(1:10)
last(1:10)
}

