% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{flatten_list}
\alias{flatten_list}
\title{Flatten a list}
\usage{
flatten_list(x, lev = 1)
}
\arguments{
\item{x}{a list}

\item{lev}{the level to which the list is to be flatten. Calculated using 
\code{link{list_depth}}}
}
\description{
Flatten a list
}
\examples{
str(x <- list(a = list(b = 1, c = list(d = 2, e = 3)), f = 4, g = list(h = list(i = 5))))
str(flatten_list(x, 1))
str(flatten_list(x, 2))
}
\keyword{internal}

