% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_phys.r
\name{front}
\alias{front}
\title{Add averaged oceonagraphic fronts to an existing plot}
\usage{
front(name = c("SACCF", "PF", "SAF", "SSTF"), col = 1:4, lwd = 2,
  rescaleLon = NULL, rescaleLat = NULL, ...)
}
\arguments{
\item{name}{The name of the front(s) to add. To choose
in \code{c('SACCF', 'PF', 'SAF', 'SSTF', 'NSTF')}}

\item{col}{The colors associated with \code{names}.}

\item{lwd}{The width of the line.}

\item{rescaleLon}{If not \code{NULL} then \code{\link{rescale}} is called
with these provided arguments.}

\item{rescaleLat}{If not \code{NULL} then \code{\link{rescale}} is called
with these provided arguments.}

\item{...}{Other atguments to be passed to \code{\link{lines}}.}
}
\description{
Add averaged oceonagraphic fronts to an existing plot
}
\details{
See the documentation of the \code{\link{sofronts}} dataset for more
information about the data.
}
\examples{
plot(c(40, 120), c(-80,-30), type = 'n', xlab = '', ylab = '')
front()
}

