% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_functionals.r
\name{\%c\%}
\alias{\%c\%}
\title{Operator for function composition}
\usage{
g \%c\% f
}
\arguments{
\item{g}{a function}

\item{f}{a function}
}
\value{
\code{g \%c\% f} returns the composed \code{g(f())}.
}
\description{
Operator for function composition
}
\examples{
f <- sum \%c\% is.na
f(c(1, NA, NA))
}
\keyword{functional}
\keyword{internal}

