% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{\%else\%}
\alias{\%else\%}
\title{Else special operator}
\usage{
val \%else\% def
}
\arguments{
\item{val}{Normal output.}

\item{def}{Default output when \code{val} is \code{FALSE}, \code{NULL} or empty.}
}
\description{
Discard first value if \code{FALSE}, \code{NULL}, empty or \code{"try-error"}
}
\examples{
"abc" \%else\% "Another value is returned"
NULL \%else\% "Another value is returned"
try(log("abc"), silent = TRUE) \%else\% "Another value is returned"
}
\keyword{internal}

