% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{\%w/i\%}
\alias{\%w/i\%}
\alias{\%w/o\%}
\title{x with(in/out) y}
\usage{
x \%w/i\% y

x \%w/o\% y
}
\arguments{
\item{x}{Vector or NULL: the values to be matched.}

\item{y}{Vector or NULL: the values to be matched against.}
}
\description{
x with(in/out) y
}
\examples{
(1:10) \%w/i\% c(3,7,12) # 3 7
(1:10) \%w/o\% c(3,7,12) # 1  2  4  5  6  8  9 10
}
\keyword{internal}

