% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{ind}
\alias{ind}
\title{Set and get the current individual}
\usage{
ind(value, cache = FALSE)
}
\arguments{
\item{value}{If provided this value becomes te current individual. If omited
the function return the last declared individual.}

\item{cache}{Should the object be copied in a cache rather than a link to
the object ?}
}
\description{
Set and get the current individual
}
\examples{
data(exses)
ind(exses)
exses$test <- "test!"
identical(ind(), exses)
}
\seealso{
\code{ind} is convenient to use with \code{\link{tdrply}}.
}

