% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{is.injective}
\alias{is.injective}
\title{Check if f: X -> Y is injective}
\usage{
is.injective(x, y = NULL)
}
\arguments{
\item{x}{x data}

\item{y}{y data}
}
\description{
f is injective: for all (x1, x2) in X^2, f(x1) = f(x2) => x1 = x2.
}
\keyword{internal}

