% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_correction.r
\name{is_dive_truncated}
\alias{is_dive_truncated}
\title{Check if a dive profile is complete or truncated}
\usage{
is_dive_truncated(x, max.diff = 5, error = FALSE)
}
\arguments{
\item{x}{The depth sequence of a dive profile.}

\item{error}{Should the function give an error if a truncated dive is
encountered ?}

\item{max_diff}{The maximum depth difference tolerated between the start and
the end of the dive before the dive is considered as truncated.}
}
\description{
Check if a dive profile is complete or truncated
}
\examples{
\dontrun{
data(exses)
tdrply(is_dive_truncated, "depth", error = TRUE, obj = exses)
}
}

