% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_phys.r
\name{isobath}
\alias{isobath}
\title{Add isobaths to an existing plot (Kerguelen area)}
\usage{
isobath(depths = -1000, lty = 2, col = "gray", plot = TRUE, ...)
}
\arguments{
\item{depths}{The isobath(s) which is(are) desired.}

\item{lty}{The line type.}

\item{col}{The line color.}

\item{plot}{Should the lines be drawn ?}

\item{...}{Other arguments to be passed to \code{\link{lines}}.}
}
\description{
Add isobaths to an existing plot (Kerguelen area)
}
\details{
Isobath at -1000 m corresponds approximatively to the Kerguelen shell 
boarder. See the documentation of the \code{\link{kerbathy}} dataset for more
information about the data.
}
\examples{
\dontrun{
# Require package 'sp' 
plot(c(68, 80), c(-54,-48), type = 'n', xlab = '', ylab = '')
isobath()
}
}

