% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{kerbathy}
\alias{kerbathy}
\title{Bathymetry of the Kerguelen area}
\format{A SpatialLinesDataFrame (S4 object of \code{sp} package)}
\usage{
data(kerbathy)
}
\description{
This dataset is a imported shapefile containing the Hi-Res. isobath of the
Kerguelen area. Available isobaths from -200 m to -4400 meters by 200 m steps.
}
\references{
\url{https://www.ga.gov.au/products/servlet/controller?event=GEOCAT_DETAILS&catno=71552}
}
\seealso{
\code{\link{isobath}}
}
\keyword{datasets}

