% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{li}
\alias{li}
\title{Linear interpolation}
\usage{
li(x, n_max = NULL)
}
\arguments{
\item{x}{A vector with missing values to interpolate.}

\item{n_max}{The maximun number of successive missing values to interpolate.}
}
\description{
Linear interpolation
}
\keyword{internal}

