% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{list_depth}
\alias{list_depth}
\title{Depth of an R object}
\usage{
list_depth(x)
}
\arguments{
\item{x}{The object to analyse.}
}
\description{
Depth of an R object
}
\details{
function \code{plotrix::maxDepth}
}
\keyword{internal}

