% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{max_residual}
\alias{max_residual}
\title{Get the maximun residual of a BSM at a given iteration}
\usage{
max_residual(x, iter = NULL, type = c("normal", "absolute"))
}
\arguments{
\item{x}{\code{bsm} object as returned by \code{\link{brokenstick}} or 
\code{\link{optBrokenstick}}.}

\item{iter}{the iteration number for which the maximum residual is to be 
returned (Ri). If \code{NULL} then \code{iter} is set to the last iteration 
in \code{x}}

\item{type}{To choose in \code{c('normal', 'absolute')}. The second choice 
returning the absolute value of the first.}
}
\description{
Get the maximun residual of a BSM at a given iteration
}
\examples{
data(exses)
dv <- tdrply(identity, 1:2, no = 100, obj = exses)[[1]]
bsm <- brokenstick(dv) 
max_residual(bsm, iter = 5)

\dontrun{
max_residual(eco.mem(bsm), iter = 5) # error
max_residual(eco.mem(bsm), iter = 4)
}
}
\keyword{brokenstick}
\keyword{internal}

