% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functionals.r
\name{min_n}
\alias{min_n}
\title{Create a function that return a result if number of valid obs >= n}
\usage{
min_n(f, n)
}
\arguments{
\item{f}{a function having a vector of data as first argument}

\item{n}{the minimum number of observation required to return a non-NA 
result. A number in \code{(0;1[} will be interpreted as a proportion.}
}
\description{
Create a function that return a result if number of valid obs >= n
}
\examples{
mean5 <- min_n(mean, 5)
mean5(c(1:4, NA))
mean5(c(1:5, NA))

mean90percent <- min_n(mean, 0.90)
mean90percent(c(1:8, NA, NA))
mean90percent(c(1:9, NA))
}
\keyword{internal}

