% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{nearly_equal}
\alias{nearly_equal}
\title{Vectorized predicate for near equality testing of numerics}
\usage{
nearly_equal(x, y)
}
\arguments{
\item{x, y}{two numeric vectors}
}
\description{
Vectorized predicate for near equality testing of numerics
}
\details{
Pairwise compariosn such as \code{`==`}.
}
\keyword{internal}

