% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_tdr.r
\name{no_interpreter}
\alias{no_interpreter}
\title{Subset dive numbers with values as using indices}
\usage{
no_interpreter(x, obj = ind())
}
\arguments{
\item{x}{a numeric or logical.}

\item{obj}{a "ses" object.}
}
\value{
vector of selected dive numbers.
}
\description{
Subset dive numbers with values as using indices
}
\details{
-0 syntax allowed.
}
\examples{
no_interpreter(NULL)    # all dive numbers returned
no_interpreter(1:10)    # returns dive numbers 1:10
no_interpreter(-(1:10)) # all dive number returned but 1:10
no_interpreter(-0)      # All dive number returned but 0
}
\keyword{internal}
\keyword{tdr_interpreter}

