% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{nstr}
\alias{nstr}
\title{nstr}
\usage{
nstr(x)
}
\arguments{
\item{x}{The object to analyse.}
}
\description{
Recursive extraction of names (such as \code{names(c(x, recursive=TRUE))}) but
stops when a subelement is atomic (avoid long long run when launched on
large object such as a TDR dataset).
}
\examples{
x <- data.frame(X=1:10, Y=10:1)
names(c(x, recursive=TRUE))
nstr(x)
}
\keyword{internal}

