% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{numIn}
\alias{numIn}
\title{Extract numbers in character strings}
\usage{
numIn(x, simplify = FALSE)
}
\arguments{
\item{x}{Atomic vector or list.}

\item{simplify}{Logical or character string. Should the result be simplified 
to a vector, matrix or higher dimensional array if possible?}
}
\description{
Extract numbers in character strings
}
\examples{
# Atomic character
x <- levels(cut(1:100, 3))
(out <- numIn(x))

# Atomic factor is coerced to character
x <- unique(cut(1:100, 3))
identical(numIn(x), out) 	# TRUE

# Works on list as well
x <- do.call(list, as.list(x))
identical(numIn(x), out) 	# TRUE

# When type is not character or factor the names are used
x <- do.call(list, as.list(1:3))
names(x) <- unique(cut(1:100, 3))
identical(numIn(x), out) 	# TRUE

# If names is NULL or empty the row.names are used instead
x <- matrix(1:6, 3)
row.names(x) <- levels(cut(1:100, 3))
is.null(names(x)) 		# TRUE
identical(numIn(x), out) 	# TRUE
}
\keyword{internal}

