% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{optBrokenstick}
\alias{optBrokenstick}
\title{Fitting automatic brokenstick models}
\usage{
optBrokenstick(x, y = NULL, threshold, cost = max_dist_cost, npmin = 2,
  npmax = Inf, start = NULL, na.action, ...)
}
\arguments{
\item{x}{The x data. Note that the \code{x} values have to be sorted. Can also 
be a list of \code{x} and \code{y} data as returned by \code{\link{xy.coords}}.}

\item{y}{The y data.}

\item{threshold}{A threshold value for the cost function to be used instead of 
the minimum. If provided the search of a local minimum in the cost function is 
abandoned.}

\item{cost}{The cost function to use. Some are included in the package such as 
\code{\link{max_dist_cost}}, \code{\link{dist_per_pt_cost}}, 
\code{\link{rss_cost}}, \code{\link{dzi_cost}} etc. Feel free to use a custom one.}

\item{npmin}{Minimun number of points.}

\item{npmax}{Maximum number of points.}

\item{start}{Some starting points to start the algorithm with.}

\item{na.action}{A function which indicates what should happen when the data 
contain \code{NAs}. The default is set by the \code{na.action} setting of 
\code{\link{options}}, and is \code{\link{na.fail}} if that is unset. The 
"factory-fresh" default is \code{\link{na.omit}}. Another possible value 
is \code{NULL}, no action. Value \code{\link{na.exclude}} can be useful.}

\item{...}{Further arguments to be passed to \code{\link{bsmfit}} 
such as \code{eco.mem}.}
}
\value{
Same as \code{\link{brokenstick}} with the value of the cost function.
}
\description{
\code{optBrokenstick} is similar to \code{\link{brokenstick}} except that a 
cost function can be used to determine the optimal number of points.
}
\examples{
data(exses)
dv <- tdrply(identity, 1:2, no = 90, obj = exses)[[1]]

bsm_6p <- brokenstick(dv, npts = 6)
plot(bsm_6p, data = TRUE)

bsm_30m <- optBrokenstick(dv, threshold = 30, cost = max_dist_cost) 
plot(bsm_30m, add = TRUE, col = 2, lty = 2, enumerate = TRUE, 
     col.pts = (bsm_30m$pts.no > 5) + 1)

bsm_5m <- optBrokenstick(dv, threshold = 5, cost = max_dist_cost) 
plot(bsm_5m, add = TRUE, col = 3, lty = 3, enumerate = TRUE, 
     col.pts = (bsm_5m$pts.no > max(bsm_30m$pts.no)) + (bsm_5m$pts.no > 5) + 1)
}
\seealso{
\code{\link{brokenstick}} and \code{\link{predict.bsm}}, \code{\link{residuals.bsm}}, 
\code{\link{update.bsm}}, \code{\link{summary.bsm}}, 
\code{\link{coef.bsm}}, \code{\link{plot.bsm}}, \code{\link{as.data.frame.bsm}}
for other functions with a S3 method for \code{bsm} objects.
}
\keyword{brokenstick}

