% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_acc_processing.r
\name{overall_DBA}
\alias{overall_DBA}
\title{Overall Dynamic Body Acceleration (ODBA)}
\usage{
overall_DBA(object)
}
\arguments{
\item{object}{A data frame or TDR table including dynamic acceleration variables 
entitled "axD", "ayD", and "azD" for X, Y, and Z axes of the accelerometer.}
}
\value{
A vector of ODBA of the same length as \code{object}.
}
\description{
Overall Dynamic Body Acceleration (ODBA)
}
\keyword{raw_processing}

