% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_behavior.r
\name{pca_count}
\alias{pca_count}
\title{Count the number of Prey Catch Attemps (PCA)}
\usage{
pca_count(x)
}
\arguments{
\item{x}{a logical vector indicating at each timestamp if it was associated 
to a PCA event.}
}
\description{
Count the number of Prey Catch Attemps (PCA)
}
\details{
A continuous succession of \code{TRUE} is considered as a single PCA.
}
\examples{
data(exses)
btt_pca <- tdrply(pca_count, "is_pca", ty = "_", obj = exses)
}

