% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{per}
\alias{per}
\title{Decompose an atomic vector to its successive values and their length.}
\usage{
per(x, idx = TRUE)
}
\arguments{
\item{x}{The atomic vector to examine.}

\item{idx}{Should the indexes (start and end) of homogeneous sequences be 
returned as well ?}
}
\value{
A data frame with values and lengths of the homogeneous sequences 
of x. The class of the column 'value' is copied from the input.
}
\description{
The reverse of 'base::rep()' function: decompose an atomic vector to its successive 
values and their length.
}
\examples{
(x <- rep(LETTERS[1:10], 10:1))
(y <- per(x))
identical(rep(y$value, y$length), x)   # TRUE
inherits(y$value, class(x))            # TRUE
}
\keyword{internal}

