% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_acc_processing.r
\name{pitch}
\alias{pitch}
\alias{roll}
\title{Attitude angles from static accelation}
\usage{
pitch(object)

roll(object)
}
\arguments{
\item{object}{A data frame or TDR table including static acceleration variables
entitled "axG", "ayG", and "azG" for X, Y, and Z axes of the accelerometer.}
}
\value{
A vector of pitch/roll of the same length as \code{object}.
}
\description{
Attitude angles from static accelation
}
\details{
For roll angle, the x axe is not necessary.
}
\keyword{raw_processing}

