% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{plot.bsm}
\alias{plot.bsm}
\title{Plot brokenstick models}
\usage{
\method{plot}{bsm}(x, type = "b", lwd = 2, ylim = rev(range(xy$y)),
  add = FALSE, col = 1, col.pts = col, enumerate = FALSE,
  data = FALSE, ...)
}
\arguments{
\item{x}{\code{bsm} object to plot, typically result from \code{\link{brokenstick}} 
or \code{\link{optBrokenstick}}.}

\item{type}{Character indicating the type of plotting; actually any of the 
types as in \code{\link{plot.default}}.}

\item{lwd}{The line width, a positive number, defaulting to 2.}

\item{ylim}{the y limits (y1, y2) of the plot. 
Here y1 > y2 and leads to a "reversed axis".}

\item{add}{If true add the plot to the already existing plot.}

\item{col}{A specification for the default plotting color.}

\item{col.pts}{Color of the breakpoints (can be of length > 1).}

\item{enumerate}{A switch to indicate if the iteration number of points should 
be added to the plot.}

\item{data}{Should the data used to fit the BSM be plotted as well ?}

\item{...}{Further graphical parameters (see \code{\link{par}}) may also be 
supplied as arguments.}
}
\description{
Plot brokenstick models
}
\details{
Require \code{eco.mem <= 4} (see \code{\link{bsmfit}}).
}
\examples{
data(exses)
dv <- tdrply(identity, 1:2, no = 100, obj= exses)[[1]]
bsm <- brokenstick(dv) 
plot(bsm, data = TRUE, enumerate = TRUE)

# Similar (but plot.tdr draws POSIXct on x axis while plot.bsm draws numerics)
plot(dv) 
plot(bsm, add = TRUE, enumerate = TRUE)
}
\seealso{
\code{\link{brokenstick}}, \code{\link{optBrokenstick}}
}
\keyword{brokenstick}
\keyword{internal}

