% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{plot.dzi}
\alias{plot.dzi}
\title{Plot method for "dzi" objects}
\usage{
plot.dzi(x, dz_col = "lightblue", dz_border = "blue", enumerate = TRUE,
  ...)
}
\arguments{
\item{x}{a \code{"dzi"} object}

\item{dz_col}{color of dive zone area}

\item{dz_border}{color of the border of the dive zone area}

\item{enumerate}{should the order of BSM break points be enumerated.}

\item{...}{Arguments to be passed to methods, such as graphical
parameters (see \code{\link{par}}).}
}
\description{
Plot method for "dzi" objects
}
\keyword{brokenstick}

