% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_plot.r
\name{plot.tdr}
\alias{plot.tdr}
\title{plot for TDR data}
\usage{
\method{plot}{tdr}(x, type = "l", ...)
}
\arguments{
\item{x}{TDR data}

\item{type}{as in \code{\link{plot}}}

\item{...}{as in \code{\link{plot}}}
}
\description{
Reverse Y axe when depth column in inputs
}
\keyword{internal}

