% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{pmean}
\alias{pmean}
\title{pmean}
\usage{
pmean(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric or logical arguments}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\description{
Returns the (parallel) average of the input values.
}
\examples{
pmean(1:3, 3:1)
pmean(1:3, 0)
}
\keyword{internal}

