% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{predict.bsm}
\alias{predict.bsm}
\title{Brokenstick model predictions}
\usage{
\method{predict}{bsm}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class inheriting from "\code{bsm}".}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}

\item{...}{other arguments.}
}
\description{
Brokenstick model predictions
}
\details{
Require \code{eco.mem <= 4} (see \code{\link{bsmfit}}). When
using \code{newdata} argument, the function returns \code{NAs} for values that
do not belong to a stick.
}
\examples{
data(exses)
dv <- tdrply(identity, 1:2, no = 100, obj = exses)[[1]]
bsm <- brokenstick(dv)
plot(depth ~ time, dv, ylim = rev(range(dv$depth)), type = 'l')
plot(bsm, add = TRUE, enumerate = TRUE)

ypred <- predict(bsm, newdata = xpred <- sample(dv$time, 10))
points(xpred, ypred, col = 3, pch = 19)
}
\seealso{
\code{\link{brokenstick}}, \code{\link{optBrokenstick}},
\code{\link{which.stick}}
}
\keyword{brokenstick}
\keyword{internal}

