% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{print.bsm}
\alias{print.bsm}
\title{Print method for bsm objects}
\usage{
\method{print}{bsm}(x, ...)
}
\arguments{
\item{x}{a bsm object}

\item{...}{for generic compatibility}
}
\description{
Print method for bsm objects
}
\keyword{brokenstick}
\keyword{internal}

