% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{print.dzi}
\alias{print.dzi}
\title{Print method for dzi objects}
\usage{
\method{print}{dzi}(x, ...)
}
\arguments{
\item{x}{a dzi object}

\item{...}{for generic compatibility}
}
\description{
Print method for dzi objects
}
\keyword{brokenstick}
\keyword{internal}

