% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_phys.r
\name{prof}
\alias{prof}
\title{Compute temperature profile from TDR recordings}
\usage{
prof(x, y = NULL, by = 5, na.rm = FALSE)
}
\arguments{
\item{x}{the \code{x} variable.}

\item{y}{The \code{y} variable.}

\item{by}{To specify how to bin the x data.}

\item{na.rm}{Should the NAs be removed from the output ?}
}
\description{
Compute temperature profile from TDR recordings
}
\examples{
data(exses)
temp_profiles <- tdrply(prof, c("depth", "temp"), by = 10, obj = exses)
plot(temp_profiles[["!_/#33"]], type = 'b')
}

