% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_acc_processing.r
\name{read.wcih}
\alias{read.wcih}
\title{Import data from Wildlife Computers ".tab" text files}
\usage{
read.wcih(x, dt = TRUE, ...)
}
\arguments{
\item{x}{filename to be imported or a TDR dataset to be formated.}

\item{dt}{if TRUE function will return a data.table object, else, a data.frame.}

\item{...}{Arguments to be passed to \code{\link{file}} such as \code{encoding}.}
}
\description{
Import data from Wildlife Computers ".tab" text files
}
\details{
Wildlife Computers > Instrument helper > Save instrument readings > R format
}
\keyword{raw_processing}

