% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{replaceMissing}
\alias{replaceMissing}
\title{Replace values in an atomic vector.}
\usage{
replaceMissing(x, na.0 = NaN, na.1 = NA)
}
\arguments{
\item{x}{The atomic vector}

\item{na.0}{The value to be replaced. Default is NaN.}

\item{na.1}{The replacement. Default is NA.}
}
\description{
Replace values in an atomic vector.
}
\examples{
x <- sample(c(1:3,NaN), 20, replace=TRUE)
x
replaceMissing(x)
}
\keyword{internal}

