% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{residuals.bsm}
\alias{residuals.bsm}
\title{Extract brokenstick model residuals}
\usage{
\method{residuals}{bsm}(object, type = c("normal", "absolute"), newdata, ...)
}
\arguments{
\item{object}{\code{bsm} object, typically result from \code{\link{brokenstick}} 
or \code{\link{optBrokenstick}}.}

\item{type}{To choose in \code{c('normal', 'absolute')}. The second choice 
returning the absolute value of the first.}

\item{newdata}{\code{x} and \code{y} data to used when interested in specific 
residual values (\code{x} in first column, \code{y} in second column).}

\item{...}{Other arguments.}
}
\value{
Returns model residuals if \code{newdata} is omited. Returns residuals 
computed from \code{newdata} otherwise. In this case, returns \code{NAs} for 
values that do not belong to a stick .
}
\description{
Extract brokenstick model residuals
}
\details{
Require \code{eco.mem <= 4} if using \code{newdata} argument but requires 
\code{eco.mem <= 2} otherwise (see \code{\link{bsmfit}}).
}
\examples{
data(exses)
dv <- tdrply(identity, 1:2, no = 400, obj= exses)[[1]]
bsm <- brokenstick(dv)
plot(residuals(bsm)) ; abline(v = bsm$pts, h = 0)
}
\seealso{
\code{\link{brokenstick}}, \code{\link{optBrokenstick}}
}
\keyword{brokenstick}
\keyword{internal}

