% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_correction.r
\name{resize}
\alias{resize}
\title{Resize an individual}
\usage{
resize(x, no_dv_min = 1, no_dv_max = max(x$delim$no_dive), rm = NULL)
}
\arguments{
\item{x}{An object of class \code{ses}.}

\item{no_dv_min}{The number of the first dive to keep.}

\item{no_dv_max}{The number of the last dive to keep.}

\item{rm}{Optional. A logical vector of length = \code{nrow(x$delim)}, where
\code{TRUE} indicate that dives are to be removed.}
}
\description{
\code{resize} cuts a \code{ses} object: given a set of dives to discard, it
remove the according rows from the TDR dataset and the statdive table. The dive
and bottom delimitation ("delim" table) is updated as well.
}
\examples{
data(exses)
range(exses$tdr$time)
range(abs(exses$delim$no_dive))
range(exses$stat$no_dive)

new.ses <- resize(exses, 50, 55)
range(new.ses$tdr$time)
range(abs(new.ses$delim$no_dive))
range(new.ses$stat$no_dive)
}
\keyword{correction}

