% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functionals.r
\name{rollapply}
\alias{rollapply}
\title{Apply a function along a vector or a list}
\usage{
rollapply(obj, fun, w = 5, ..., wty = c("moving", "fixed"),
  simplify = TRUE, aty = c("single", "multiple"),
  nas = pmatch(match.arg(aty), "single", 0))
}
\arguments{
\item{obj}{A list or an atomic vector.}

\item{fun}{A function to be applied.}

\item{w}{Width of the window.}

\item{...}{Optional arguments to \code{fun}.}

\item{wty}{Window type: to choose between \code{c('moving', 'fixed'))}.}

\item{simplify}{Logical or character string. Should the result be simplified 
to a vector, matrix or higher dimensional array if possible?}

\item{aty}{Argument type: How the elements of \code{obj} should be given to \code{fun}. 
\code{'single'} if they should be regouped in one argument. \code{'multiple'} 
if they should be considered as separated arguments. See the last example.}

\item{nas}{Should NAs be added to the result to match the length of the input.}
}
\description{
Apply a function along a vector or a list
}
\examples{
x <- seq(1, 3, length = 150) + rt(150, df = 2) / 3
plot(x)
av  <- rollapply(x, mean, 15)
lines(av, lwd = 2)
err <- rollapply(x, sd, 15)
lines(av + err, lwd = 2, lty = 2) ; lines(av - err, lwd = 2, lty = 2)

# 'wty' argument
lines(rollapply(x, mean, wty = 'f'), type = 's', col = 'red')

# 'aty' argument
x <- 2^(0:5)
out <- rollapply(x, seq, 2, aty = 'm')
names(out) <- rollapply(x, paste, 2, sep = ':', aty = 'm')
out
}
\keyword{internal}

