% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_biopic.r
\name{SI_light}
\alias{SI_light}
\alias{WC_light}
\title{Convert Wildlife Computers light values from/to linear W/cm^2 units}
\usage{
SI_light(x)

WC_light(x)
}
\arguments{
\item{x}{raw sensor readings}

\item{x}{transformed sensor readings}
}
\description{
Use the equation provided by the manufacturer, Wildlife Computers.
}
\keyword{light}

