% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_acc_processing.r
\name{static_acc}
\alias{static_acc}
\title{Static acceleration}
\usage{
static_acc(x, fs = 16, fc = 0.2)
}
\arguments{
\item{x}{3 axes acceleration table with time in the first column and acceleration 
axes in the following columns. Variables must be entitled "time" for time, 
"ax", "ay", and "az" for x, y and z accelerometer axes.}

\item{fs}{sampling frequency of the input data (Hz).}

\item{fc}{Cut-off frequency for the butterworth low pass filter (Hz)}
}
\value{
returns a data.frame with time, and X, Y and Z static accelearyion at 1 Hz.
}
\description{
The raw acceleration is first filtered using a low pass butterworth filter. 
The extracted signal is then scaled so that the norm of the the vector G is 1 at 
each second.
}
\details{
This filtered acceleration can be used to compute pitch and roll angles
}
\keyword{raw_processing}

