% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_behavior.r
\name{straightness}
\alias{sinuosity}
\alias{straightness}
\title{Compute straightness index}
\usage{
straightness(x, y = NULL, npts = 3)

sinuosity(x, y = NULL, npts = 3)
}
\arguments{
\item{x}{The x data (time).}

\item{y}{The y data (depth).}

\item{npts}{The number of points to use. 2 is the minimum (from the start
to the end of data). Each new point adds a step (using the \code{\link{brokenstick}}
algorithm) which is taken into account when computing \code{L}.}
}
\value{
\code{straightness} returns a number between 0 (maximum sinuosity)
to 1 (maximum straightness). \code{sinuosity} is equivalent to
\code{1 / straightness}.
}
\description{
\code{straightness} compute a straightness index between 0 and 1 by making the
ration \code{L / l} where \code{L} is the cumulated distance between
brokenstick points and and where \code{l} the cumulated distance between
each \code{y} data points.
}
\examples{
data(exses)
ind(exses)
sunflowerplot(tdrply(straightness, cl = 1:2, ty = '_'), exses$stat$pca)
}
\keyword{behavior}

