% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{summary.bsm}
\alias{summary.bsm}
\title{Print summary of a brokenstick model}
\usage{
\method{summary}{bsm}(object, ...)
}
\arguments{
\item{object}{a brokenstick model}

\item{...}{for method consistency}
}
\description{
Print summary of a brokenstick model
}
\examples{
data(exses)
dv <- tdrply(identity, 1:2, no = 53, obj = exses)[[1]]
summary(brokenstick(dv))
summary(brokenstick(dv, npts = 12))
}
\keyword{brokenstick}
\keyword{internal}

