% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_phys.r
\name{sun_position}
\alias{sun_position}
\title{Compute sun azimuth and elevation given a location and a date/time}
\usage{
sun_position(time, lat, lon)
}
\arguments{
\item{time}{a POSIXlt, POSIXct or data.frame. If a data frame order of 
columns must be from Year to Second (6 columns).}

\item{lat}{numeric vector of latitudes.}

\item{lon}{numeric vector of longitudes.}
}
\description{
Compute sun azimuth and elevation given a location and a date/time
}
\details{
Based on: 
\url{http://stackoverflow.com/questions/8708048/position-of-the-sun-given-time-of-Day-latitude-and-longitude.}. 
Update version: Jan 6 '12 at 21:40 by "Josh O'Brien".
}
\seealso{
\code{\link{is_day}}
}
\keyword{internal}

