% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_phys.r
\name{sunPos}
\alias{sunPos}
\title{sunPos}
\usage{
sunPos(Year, Month, Day, Hour, Minute, Second, Lat, Lon)
}
\description{
Compute sun azimuth and elevation given a location and a date/time. Algorithm taken from: 
\url{http://stackoverflow.com/questions/8708048/position-of-the-sun-given-time-of-Day-latitude-and-longitude.}. 
Update version: Jan 6 '12 at 21:40 by "Josh O'Brien".
}
\seealso{
\code{\link{sun_position}}
}
\keyword{internal}

