% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_behavior.r
\name{symmetry_index}
\alias{symmetry_index}
\title{Dive/bottom symmetry index}
\usage{
symmetry_index(x = ind(), type = c("_", "!_/"), time_col = 1,
  depth_col = 2)
}
\arguments{
\item{x}{input data, a \code{ses} object.}

\item{type}{Should the index be computed on the bottom phase only 
(\code{type = "_"}) or on the complete dive profile (\code{type = "!_/"})}

\item{time_col}{Character or numeric giving the column of the TDR table that 
stores the timestamps.}

\item{depth_col}{Character or numeric giving the column of the TDR table that 
stores the depth sequence}
}
\description{
Based on the moment where animal reaches the dive maximum depth.
}
\details{
The index ranges from 0 (skewed to the left) to 1 (skewed to the right).
}
\examples{
data(exses)
exses$stat$btt.sym <- symmetry_index(exses)
exses$stat$dv.sym <- symmetry_index(exses, type = "!_/")

plot(exses$stat[ , c("time", "btt.sym", "dv.sym")])
}
\references{
Halsey, L.G., Bost, C.-A., Handrich, Y. (2007) A thorough and 
quantified method for classifying seabird diving behaviour. 
Polar Biology, 30, 991-1004.
}
\keyword{behavior}

