% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tdr.r
\name{tdrexpand}
\alias{tdrexpand}
\title{Expand a summary variable to the length of TDR data}
\usage{
tdrexpand(x, ty = "!_/", na_value = NA, obj = ind(), ...)
}
\arguments{
\item{x}{variable to expand.}

\item{ty}{a kind of regular expression that explains how to 
duplicate x values along TDR rows.}

\item{na_value}{the NA value to be used (0 is can be useful).}

\item{obj}{a "ses" object}

\item{...}{arguments to be passed to \code{merge.data.table}.}
}
\value{
a vector of the same length as the number of rows in the tdr table of 
"obj".
}
\description{
Expand a summary variable to the length of TDR data
}
\examples{
data(exses)
ind(exses)
exses$tdr$no_btt <- tdrexpand(exses$stat$no_dive, "_")
exses$tdr$no_dive  <- tdrexpand(exses$stat$no_dive, "!_/")
}
\seealso{
\code{\link{tdrply}}
}

