% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_delim.r
\name{time_reso}
\alias{time_reso}
\title{Sampling rate of a TDR dataset}
\usage{
time_reso(x, type = c("period", "frequence"))
}
\arguments{
\item{x}{The time sequence.}

\item{type}{Should the result be returned in seconds ("period") or hertz ("frequence") ?}
}
\description{
Sampling rate of a TDR dataset
}
\details{
Sampling interval taken as the greatest common divisor of observed intervals.
}
\examples{
time_reso(seq(0, 10, by = 1/0.5), type = "freq")
time_reso(seq(0, 10, by = 1),     type = "freq")
time_reso(seq(0, 10, by = 1/2),   type = "freq")
time_reso(seq(0, 10, by = 1/5),   type = "freq")
time_reso(seq(0, 10, by = 1/16),  type = "freq")

# Please notice that exact equality may not be always true 
# due to realities of computer arithmetic
\dontrun{
# Assuming a sufficient number of decimales...
repeat {
  f <- runif(1, 0, 100)
  tst <- seq(0, 10, by = 1/f)
  if (length(tst) > 1) break
}
time_reso(tst, type = "f") == f          # is TRUE in most cases
all.equal(time_reso(tst, type = "f"), f) # is TRUE
}
}
\author{
Yves Le Bras, Simon Wotherspoon
}
\seealso{
\code{\link{all.equal}}
}
\keyword{internal}

