% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_plot.r
\name{transp}
\alias{darker}
\alias{lighter}
\alias{transp}
\title{Color helpers}
\usage{
transp(col, factor = 0.5)

lighter(col, factor = 0.5)

darker(col, factor = 0.5)
}
\arguments{
\item{col}{A vector of colors to transform.}

\item{factor}{A value ranging between 0 and 1 decribing the strenght of
the effet.}
}
\value{
\code{transparent} returns a transparent version of the input colors
with an alpha equal to the factor argument.

\code{lighter} returns a brightened version of the input colors.

\code{darker} returns a darkened version of the input colors.
}
\description{
Set alpha transparency parameter of input colors.
}
\keyword{internal}

