% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_tdr.r
\name{ty_delim}
\alias{ty_delim}
\title{Extract indices from "delim" table using a regexp-like syntax}
\usage{
ty_delim(ty = "!_/", no = NULL, obj = ind(), no_match = "ignore")
}
\arguments{
\item{ty}{a kind of regular expression to subet diving periods of a TDR dataset.}

\item{no}{the dive numbers to extract.}

\item{obj}{a "ses" object.}

\item{...}{arguments to be passed to \code{\link{check_nomtch}}}
}
\description{
Extract indices from "delim" table using a regexp-like syntax
}
\examples{
data(exses)
ind(exses)

str(tmp <- ty_delim('_', no = 50:52))
str(tmp <- ty_delim('!_/', no = 50:52))

str(tmp <- ty_delim('-&!&_&/&~', no = 50:52))
str(tmp <- ty_delim('-!&/~', no = 50:52))

str(tmp <- ty_delim('-!|/~', no = 50:52))
str(tmp <- ty_delim('-(!|/)~', no = 50:52))

str(tmp <- ty_delim('-(!|/)&~', no = 50:52))

str(tmp <- ty_delim("-&(__)|(/~)&//", no = 50:52))
}
\keyword{internal}
\keyword{tdr_extract}

