% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{upd_warn}
\alias{upd_warn}
\title{Update warning column in delim table}
\usage{
upd_warn(x, msg)
}
\arguments{
\item{x}{Subset of warn column}

\item{msg}{Message to append}
}
\description{
Update warning column in delim table
}
\keyword{internal}

