% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{update.bsm}
\alias{update.bsm}
\title{Update and Re-fit a brokenstick model}
\usage{
\method{update}{bsm}(object, npts, allow.dup = FALSE, ...)
}
\arguments{
\item{object}{Object of class inheriting from "\code{bsm}".}

\item{npts}{The number of points for the brokenstick model to fit. See
\code{\link{optBrokenstick}} for a version of \code{brokenstick} which figures
out this number of points automatically.}

\item{allow.dup}{Should the duplicated points be added anyway.}

\item{...}{Further arguments to be passed to \code{\link{bsmfit}}
such as \code{eco.mem}.}
}
\description{
Update and Re-fit a brokenstick model
}
\examples{
data(exses)
dv <- tdrply(identity, 1:2, no = 100, obj = exses)[[1]]
bsm <- brokenstick(dv)
length(bsm$pts)
length(update(bsm, 5)$pts)
length(update(bsm, 7)$pts)
}
\keyword{brokenstick}
\keyword{internal}

