% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_acc_processing.r
\name{vectorial_DBA}
\alias{vectorial_DBA}
\title{Vectorial Dynamic Body Acceleration (VeDBA)}
\usage{
vectorial_DBA(object)
}
\arguments{
\item{object}{A data frame or TDR table including dynamic acceleration variables
entitled "axD", "ayD", and "azD" for X, Y, and Z axes of the accelerometer.}
}
\value{
A vector of VeDBA of the same length as \code{object}.
}
\description{
Vectorial Dynamic Body Acceleration (VeDBA)
}
\keyword{raw_processing}

