% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{which.bw}
\alias{which.bw}
\title{Match values against a data.frame with start and end values}
\usage{
which.bw(x, ref)
}
\arguments{
\item{x}{the values to be matched against \code{ref}}

\item{ref}{a data.frame with start values in the first column end values in 
the second column and an optional id number in the third column.}
}
\value{
for each \code{x} value, the row number of \code{ref} where \code{x} 
lies between start and end values. If \code{ref} has a third column (an id) 
its value is returned instead of the row number. When x value matches a start 
and a end value the priority is given to the start.
}
\description{
Match values against a data.frame with start and end values
}
\keyword{internal}

