% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{which.dive}
\alias{which.dive}
\title{Find to which specific dive/surface a instant belongs to}
\usage{
which.dive(x, object = ind())
}
\arguments{
\item{x}{The time (format \code{POSIXct}) or a integer giving the row number.}

\item{object}{A \code{ses} object such as returned by \code{\link{as.ses}}.}
}
\description{
Find to which specific dive/surface a instant belongs to
}

