% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_pkgtools.r
\name{which.row}
\alias{which.row}
\title{from time stamp to row number}
\usage{
which.row(x, obj = ind())
}
\arguments{
\item{x}{a POSIXct vector}
}
\description{
from time stamp to row number
}
\examples{
data(exses)
ind(exses)
x <- sample(1:nrow(exses$tdr), 100)
identical(which.row(exses$tdr$time[x]), x)
}
\keyword{internal}

