% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_behavior.r
\name{wiggles}
\alias{wiggles}
\title{Count the number of wiggle in a dive}
\usage{
wiggles(x, y = NULL, thres.depth = 2, thres.dur = c(5, 300),
  plt = FALSE)
}
\arguments{
\item{x}{The x data. A subset of a TDR dataset time sequence. Can also
be a list of \code{x} and \code{y} (processed by \code{\link{xy.coords}}).}

\item{y}{The y data. A subset of a TDR dataset depth sequence, processed by
\code{\link{xy.coords}}).}

\item{thres.depth}{minimum depth difference.}

\item{thres.dur}{minimum time difference.}

\item{plt}{Should graphics about processing be plotted ?}
}
\description{
\code{wiggles} ...
}
\examples{
data(exses)
sunflowerplot(tdrply(wiggles, ty = '_', obj = exses), exses$stat$pca)
}
\keyword{behavior}

