% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{as.bsm}
\alias{as.bsm}
\alias{is.bsm}
\title{General S3 utils for bsm objects}
\usage{
as.bsm(pts.x, pts.y = NULL, ...)

is.bsm(x)
}
\arguments{
\item{pts.x}{x coordinates of breakpoints. If \code{pts.x} is a list then 
it is interpreted as being a \code{"bsm"} object and is returned as is with 
updated class.}

\item{pts.y}{y coordinates of breakpoints}

\item{...}{other BSM slots such as \code{"data"}, \code{"fitted"} or 
\code{"residuals"}. See details of slots in \code{\link{brokenstick}}.}

\item{x}{an object to test.}
}
\description{
General S3 utils for bsm objects
}
\examples{
data(exses)
dv <- tdrply(identity, 1:2, no = 400, obj = exses)[[1]]
bsm <- brokenstick(dv) 

as.bsm(bsm$pts.x, bsm$pts.y)
as.bsm(bsm$pts.x, bsm$pts.y, residuals = "dummy")
as.bsm(bsm)
is.bsm(bsm)
}
\keyword{brokenstick}
\keyword{internal}

