% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_delim.r
\name{bottom_delim}
\alias{bottom_delim}
\title{Find the dives' bottom start and end}
\usage{
bottom_delim(obj, nms = c("time", "depth"), thres.spd = 0.9, w = 12,
  min.depth = 0.4, ...)
}
\arguments{
\item{obj}{An object of class \code{'ses'}, \code{'statdives'} or \code{'tdr'}.}

\item{nms}{The names of the variables to use. Depends on the method.}

\item{thres.spd}{The vertical speed threshold.}

\item{w}{The window width for rolling mean applied to vertical speed.}

\item{min.depth}{The minimum depth (reported to the dive maximun depth)
allowed for bottom limits.}

\item{...}{Other arguments to be passed to \code{\link{dive_delim}}.}
}
\description{
Find the dives' bottom start and end
}
\examples{
data(exses)
dvs <- bottom_delim(exses$tdr)
}
\seealso{
\code{\link{dive_delim}}
}

