% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_delim.r
\name{drift_stat}
\alias{drift_stat}
\title{Use brokensticks to identify potential drifts}
\usage{
drift_stat(object, thres.bsm = 5, bsm = NULL)
}
\arguments{
\item{object}{a \code{SES} object including a TDR dataset, a "dives statistics"
table and a "delim" table.}

\item{thres.bsm}{a threshold to use with \code{\link{max_dist_cost}} when
running \code{\link{optBrokenstick}}.}

\item{bsm}{A list of brokenstick models to use directly instead of computing a
new one with \code{thres.bsm} as parameter.}
}
\value{
A data frame of all brokenstick segments with start and end index of
the segment, no_seg_tot and ID number of the segement, no_seg_dive the number
of the segment within its dive, no_dive the ID of the dive conataining the segment,
time the starting time of segement, drift_rate the vertical speed of the segment,
dur the duration of the segment, min_depth and max_depth the minimum and maximum
depths of the segment, acceleration statistics (if acceleration data available)
average roll, pitch and swimming effort and PCA rate, the mean squared residuals
of the segment.
}
\description{
\code{drift_stat} takes an object and return a data frame of all the brokenstick
segments along the trip with statistics such as their vertical speed (average
and standard deviation), their duration and so on. This data frame is to be
filtered to identify drift segments.
}
\details{
To get average pitch and roll angles (degrees), the TDR table must include
static acceleration with variable names "axG", "ayG" and "azG". Similarly, for
swimming effort and PCA rate, the TDR table must have columns "swm_eff" (numeric)
and "is_pca" (logical). Use functions \code{\link{static_acc}},
\code{\link{swimming_effort}} and \code{\link{prey_catch_attempts}} to compute
these variables from the raw acceleration readings.
}
\examples{
data(exses)
tab <- drift_stat(exses)

\dontrun{
require(manipulate)
manipulate(
{
tab$day <- floorPOSIXct(tab$time)
 cnd <- tab$dur >= dur * 60 &
        abs(tab$drift_rate) <= rate &
        tab$swm_eff <= swm_eff &
        tab$sd_res <= sd_resid
 cnd <- "if"(no_pca, cnd & tab$pca_rate == 0, cnd)
 belly_up <- abs(tab$roll) >= 90
 if (up_only) cnd <- cnd & belly_up
 na_cnd <- is.na(cnd)
 cnd <- ifelse(is.na(cnd), TRUE, cnd)
 pts_col <- ifelse(belly_up[cnd], "red", "black")
 pts_col <- ifelse(na_cnd, "blue", pts_col)
 pts_shape <- ifelse(na_cnd, 20, 1)
 plot(drift_rate ~ time, tab[cnd, ], col = pts_col, pch = pts_shape,
      xlim = range(tab$time), ylim = c(-3, 3))
 abline(h = 0, col = 'gray')
},
 dur = slider(0, 30, initial = 1.5, step = 0.25, label = "Duration (minutes)"),
 rate = slider(0, 3, initial = 3, step = 0.05, label = "Max. vertical speed (m/s)"),
 swm_eff = slider(0, 0.5, initial = 0.1, step = .01, label = "Max. swimming effort (m/s³)"),
 sd_resid = slider(0, 10, initial = 3, step = 0.1, label = "Max residual sd (m/s)"),
 up_only = checkbox(FALSE, label = "Show belly up only"),
 no_pca = checkbox(FALSE, label = "Show no PCA only")
)

# Drift dive example
tdrply(plot, 1:2, no = 300, type = 'l', obj = exses)
}
}
\keyword{drift}

