% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tdr.r
\name{elt_delim}
\alias{elt_delim}
\title{Extract indices corresponding to a symbol from a "delim" table}
\usage{
elt_delim(elt, no, delim)
}
\arguments{
\item{elt}{A single character to match against delim table.}

\item{no}{dives numbers to extract.}

\item{delim}{a delim table (delim slot of a "ses" object).}
}
\value{
a data frame with row number of sart/end of the desired periods.
}
\description{
Extract indices corresponding to a symbol from a "delim" table
}
\examples{
data(exses)
elt_delim('-', 60:62, exses$delim)
elt_delim("/", no = c(62, 62:60), delim = exses$delim)
}
\keyword{internal}
\keyword{tdr_extract}

