% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_brokenstick.r
\name{max_dist_cost}
\alias{dist_per_pt_cost}
\alias{max_dist_cost}
\title{Cost functions for automatic brokenstick models.}
\usage{
max_dist_cost(object)

dist_per_pt_cost(object)
}
\arguments{
\item{object}{Object of class inheriting from "\code{bsm}".}
}
\value{
A statistic to minimize.
}
\description{
Given a model and data the function returns a single value which is used to
determine if the number of points is optimized: the cost function value being
minimized by the optimal set of parameters.
}
\details{
\code{max_dist_cost} In this function the statistic is the maximun distance between an
observation and its fitted value. Hence the function constantly decrease with
increasing number of point in the model and a threshold must be provided along
with this function to avoid infinite looping. Yet, the pros of this cost function
is a threshold that is simple to determine and to interpret.

\code{dist_per_pt_cost} In this function the statistic is the
average distance between observations and fitted values divided by the number of
points used by the model. This function has does rech a minimun value but
generally for high numbers of points.
}
\seealso{
\code{\link{optBrokenstick}}
}
\keyword{brokenstick}
\keyword{internal}

