% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tdr.r
\name{no_interpreter}
\alias{no_interpreter}
\title{Subset dive numbers with values as using indices}
\usage{
no_interpreter(x, obj = ind())
}
\arguments{
\item{x}{a numeric or logical.}

\item{obj}{a "ses" object.}
}
\value{
vector of selected dive numbers.
}
\description{
Subset dive numbers with values as using indices
}
\examples{
data(exses)
ind(exses)
no_interpreter(NULL)    # all dive numbers returned

# Define default values using "no_dive" slot of the ses object
exses$no_dive <- 1:15
no_interpreter(NULL)

no_interpreter(1:10)    # returns dive numbers 1:10
no_interpreter(-(1:10)) # all dive number returned but 1:10
}
\keyword{internal}
\keyword{tdr_interpreter}

