% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.r
\name{plot.tdr}
\alias{plot.tdr}
\title{plot for TDR data}
\usage{
\method{plot}{tdr}(x, type = "l", ...)
}
\arguments{
\item{x}{TDR data}

\item{type}{as in \code{\link{plot}}}

\item{...}{as in \code{\link{plot}}}
}
\description{
When depth in input set in as Y and reverse axe. Use third column as a color 
variable.
}
\examples{
data(exses)
tdrply(plot, c("time", "depth", "is_pca"), no = 65, obj = exses)

exses$tdr$no_btt <- tdrexpand(exses$stat$no_dive, ty = "_", obj = exses)
exses$tdr$is_btt <- !is.na(exses$tdr$no_btt)
tdrply(plot, c("time", "depth", "is_btt"), no = 65, obj = exses)
tdrply(plot, c("depth", "temp", "is_btt"), no = 65, obj = exses)
}
\keyword{internal}

