% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_acc_processing.r
\name{prey_catch_attempts}
\alias{prey_catch_attempts}
\title{Identify Prey Catch attempts}
\usage{
prey_catch_attempts(x, fs = 16, fc = 2.64)
}
\arguments{
\item{x}{3 axes acceleration table with time in the first column and acceleration 
axes in the following columns. Variables must be entitled "time" for time, 
"ax", "ay", and "az" for x, y and z accelerometer axes.}

\item{fs}{sampling frequency of the input data (Hz).}

\item{fc}{Cut-off frequency for the butterworth high pass filter (Hz)}
}
\value{
returns a logical vector of prey catch attempts at 1 Hz frequency. 
Value is TRUE if the record belong to prey catch attempt FALSE otherwise.
}
\description{
Identify Prey Catch attempts
}
\keyword{raw_processing}

