% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funs_delim.r
\name{time_reso}
\alias{time_reso}
\title{Sampling rate of a TDR dataset}
\usage{
time_reso(x, n = 10, type = c("period", "frequence"))
}
\arguments{
\item{x}{The time sequence.}

\item{n}{the number of replicates to do.}

\item{type}{Should the result be returned in seconds ("period") or hertz ("frequence") ?}
}
\description{
Sampling rate of a TDR dataset
}
\keyword{internal}

