% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_delim.r
\name{time_reso}
\alias{time_reso}
\title{Sampling rate of a TDR dataset}
\usage{
time_reso(x, type = c("period", "frequence"))
}
\arguments{
\item{x}{The time sequence.}

\item{type}{Should the result be returned in seconds ("period") or hertz ("frequence") ?}
}
\description{
Sampling rate of a TDR dataset
}
\details{
Sampling interval taken as the greatest common divisor of observed intervals.
}
\author{
Simon Wotherspoon
}
\keyword{internal}

